requirements.txt
pandas==2.2.1
reportlab==4.1.0
Pillow==10.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
PyPDF2==3.0.1
boto3==1.34.69
botocore==1.34.69 